Fix: CustomBottomNavBar and related functions moved to separeted file
