+++ uikit/src/main/java/care/intouch/uikit/ui/navigation/CustomBottomNavBar.kt
@@ -6,13 +6,11 @@ import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
@@ -28,7 +26,6 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -36,7 +33,6 @@ import androidx.constraintlayout.compose.ConstraintLayout
 import care.intouch.uikit.R
 import care.intouch.uikit.theme.InTouchTheme
 
-// UI elements
 @Composable
 fun NavBottomBarPlusButton(
     modifier: Modifier = Modifier,
@@ -68,45 +64,6 @@ fun NavBottomBarPlusButtonPreview() {
     InTouchTheme {
         NavBottomBarPlusButton(
             onClick = {}
-        )
-    }
-}
-
-@Composable
-fun TopBarArcButton(
-    onClick: () -> Unit,
-    enabled: Boolean,
-    modifier: Modifier = Modifier
-) {
-    Box(
-        modifier = modifier
-            .clickable(
-                indication = null,
-                interactionSource = remember { MutableInteractionSource() },
-                onClick = { onClick.invoke() }
-            ),
-        contentAlignment = Alignment.Center
-    ) {
-       Icon(
-           painter = painterResource(id = R.drawable.arc_rectangle),
-           contentDescription = null,
-           tint = if (enabled) InTouchTheme.colors.accentColorGreen else InTouchTheme.colors.accentColorGreen50
-       )
-        Icon(
-            painter = painterResource(id = R.drawable.icon_close),
-            contentDescription = null,
-            tint = InTouchTheme.colors.inputColor
-        )
-    }
-}
-
-@Composable
-@Preview(showBackground = true)
-fun TopBarArcButtonPreview() {
-    InTouchTheme {
-        TopBarArcButton(
-            onClick = {},
-            enabled = false
         )
     }
 }
@@ -157,96 +114,6 @@ fun NavBottomComplexElementPreview() {
             text = "Home",
             painter = painterResource(id = R.drawable.icon_home),
             focusTint = InTouchTheme.colors.mainColorGreen,
-        )
-    }
-}
-
-// Navigation UI
-@Composable
-fun CustomTopBar(
-    onBackArrowClick: () -> Unit,
-    onCloseButtonClick: () -> Unit,
-    title: String,
-    enabledArcButton: Boolean,
-    addBackArrowButton: Boolean,
-    addCloseButton: Boolean
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .wrapContentHeight()
-            .background(InTouchTheme.colors.inputColor),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Center
-    ) {
-
-        if (addBackArrowButton) {
-            Box(
-                modifier = Modifier
-                    .padding(start = 28.dp)
-                    .size(44.dp)
-            ) {
-                Icon(
-                    painter = painterResource(id = R.drawable.icon_arrow_left),
-                    contentDescription = null,
-                    tint = InTouchTheme.colors.mainColorGreen,
-                    modifier = Modifier
-                        .align(Alignment.CenterStart)
-                        .clickable(
-                            indication = null,
-                            interactionSource = remember { MutableInteractionSource() },
-                            onClick = { onBackArrowClick.invoke() }
-                        ),
-                )
-            }
-        } else {
-            Box(
-                modifier = Modifier
-                    .padding(start = 28.dp)
-                    .size(44.dp)
-                    .background(Color.Transparent)
-            )
-        }
-
-        Text(
-            text = title,
-            style = InTouchTheme.typography.titleLargeTypography,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-            color = InTouchTheme.colors.textColorBlue,
-            modifier = Modifier
-                .weight(1f),
-            textAlign = TextAlign.Center
-        )
-
-        if (addCloseButton) {
-            TopBarArcButton(
-                onClick = onCloseButtonClick,
-                enabled = enabledArcButton,
-                modifier = Modifier.padding(end = 28.dp)
-            )
-        } else {
-            Box(
-                modifier = Modifier
-                    .padding(end = 28.dp)
-                    .size(47.dp)
-                    .background(Color.Transparent)
-            )
-        }
-    }
-}
-
-@Composable
-@Preview(showBackground = true)
-fun CustomTopBarPreview() {
-    InTouchTheme {
-        CustomTopBar(
-            onBackArrowClick = { /*TODO*/ },
-            onCloseButtonClick = { /*TODO*/ },
-            title = "Title Large",
-            enabledArcButton = false,
-            addBackArrowButton = true,
-            addCloseButton = true
         )
     }
 }
@@ -311,7 +178,7 @@ fun CustomBottomNavBar(
                     start.linkTo(parent.start)
                 }
         )
-        
+
         NavBottomComplexElement(
             onClick = {
                 selectedElementId = ElementId.ADDITIONAL_ID
@@ -356,7 +223,7 @@ fun CustomBottomNavBar(
                     end.linkTo(plusTag.start)
                 }
         )
-        
+
         NavBottomComplexElement(
             onClick = {
                 selectedElementId = ElementId.MY_PLAN_ID
@@ -403,4 +270,3 @@ fun CustomBottomNavBarPreview() {
 enum class ElementId {
     HOME_ID, MY_PROGRESS_ID, PLUS_ID, MY_PLAN_ID, ADDITIONAL_ID;
 }
-
